rework!: pixel shaders (#1)
this is the beginning of a reimagining. at the moment, only works on
macOS using Metal for the graphics backend. the one feature
implemented is the creation of (essentially fairly arbitrary) pixel
shaders, which must be written in MSL, are aware of the mouse, and
allow a user-defined buffer input, mouse state, and a feedback-capable
texture.

the pixel shader vertex function must be named "vertexFn" and the
fragment function must be named "fragmentFn". the current frame size
and frame number are available as `buffer(0)` and should be declared
in the shader code as

```metal
struct frameData
{
  float w;
  float h;
  uint32_t frame;
};
```

the mouse data is available as `buffer(1)` and should be declared in
the shader code as

```metal
struct mouseData
{
  float x;
  float y;
  bool left;
  bool right;
  bool other;
  bool present;
};
```

NB: currently mouse hovering is broken (i need to add an
NSTrackingArea to each PrismView that asks for it), so to get mouse
movement to register it is necessary to click and/or drag.

user data is available as `buffer(2)`, and is untested. obviously it
is your responsibility to make sure that the data you pass in makes
sense with the data type you declare in the shader code.

a read/write capable texture of the same size as the entire frame is
available as `texture(0)`. here is an example of accessing it every
frame and writing to it every tenth frame.

```metal
struct Frame
{
  float width;
  float height;
  uint32_t frame;
};

vertex float4 vertexFn(uint vertexID [[vertex_id]],
  constant float2 *positions [[buffer(0)]])
{
  float4 out = float4(0.0, 0.0, 0.0, 1.0);
  out.xy = positions[vertexID].xy;
  return out;
}

fragment float4 fragmentFn(float4 in [[position]],
  texture2d<float, access::read_write> txt [[raster_order_group(0),
  texture(0)]], constant Frame &frame [[buffer(0)]])
{
  float2 size = float2(frame.width, frame.height);
  uint32_t frame_nr = frame.frame;
  // normalized coordinates
  float2 pixel_coord = (2.0 * in.xy - size) / size.y;
  ushort2 texcoord = (ushort2)(in.xy);

  float4 txtr_val = txt.read(texcoord);

  float4 out = float4(0.0, 0.0, 0.0, 1.0);
  // do stuff here

  if ((frame_nr % 10) == 0)
  {
    txt.write(out, texcoord);
  }
  return out;
}
```

aside from pixel shaders, supports basic window creation and starting
and stopping the app.